requirements.txt
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.2
xgboost==2.0.3
statsmodels==0.14.2
plotly==5.21.0

streamlit==1.35.0

SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1

jupyterlab==4.2.4
ipykernel==6.29.4
